backend/requirements.txt
flask
flask-cors
scikit-learn
joblib
pandas
numpy
google-generativeai
lightgbm
